Live generator docs (#4339)
* Plural and abbreviation typos

* Redundant `/lib`, clearer in the listing so leave it there

* Bare generator is an error, not help

* Group argument descriptions

Argument descriptions are scattered, group them and include their
respective explanations.  (An alternative would be to tersely list the
arguments, and follow the entire list with explanations.)

* Mention these are modules, clarifies subsequent namespace discussion

* Consistent example arguments; clarify namespace `--web` option

Use the same arguments in each example use of the generator to simplify
for the reader.

Expand the explanation of `--web` with explicit module names to be clear
it is unrelated to the context module (which was also `Sales`).

Remove mention of unrelated view.