Copy lib folder before compiling assets.
I know that it is going to compile the assets again if we change the lib folder. But, when working with Tailwind(very popular) it needs to see the classes in the live view components before purging it on production.  I see it is a common problem with people trying to deploy tailwind with Live view to containers. https://pragmaticstudio.com/tutorials/adding-tailwind-css-to-phoenix