Add `phx.gen.socket` to generate the socket file
The socket file was removed from the installer, so this task is needed
when the user wants to work with websockets outside LiveView.

This commit also changes the `mix phx.gen.channel` task to include
instructions regarding this new task when the default socket -
`user_socket.ex` does not exist under `/channels`.