Build CommonJS module along with ESM and IIFE
esbuild can build CJS modules as well as ESM, and for the people using
build tools / or testing tools (such as Jest) that do not yet support
ESM modules, it is useful to have fallback when ESM is not availalbe.

This PR includes command to build CJS module along with currently
supported types, and adds entry "main" in package.json to point towards
that file.

This has been tested with Jest, which currently has very poor support
for ESM modules and allows for including "phoenix" in tests and also
stubbing it if needed.