Fix broken `check_origin: :conn`
In the internal implementation, the related function is:

```elixir
defp origin_allowed?(:conn, uri, _endpoint, %Plug.Conn{} = conn)
```

`uri` is a struct which is generated by `URI.parse/1`, the type of `scheme`
field is a string.

`conn` is a `%Plug.Conn{}`, the type of `scheme` field is an atom.

Because of that, this check will always failed.

* * *

This commit fixes above issue, and fix related tests.