Add ecto.seed alias to generator
I create an `ecto.seed` alias in every app that I use, I think it is very
useful to make it available in the aliases because a lot of the time we
want to add more seeding data without necessary running any migration.

It is also much more easy to type.