Prioritize user `:http` configuration for ranch
This puts an application's `:http` configuration for their Endpoint
first in the list to be sent into the ranch supervisor. When running on
a newer commit of ranch, this allows an application to specific
`:inet_backend` for the TCP sockets spawned by ranch.

Without this change, the application configuration is always _after_
port and IP. Since `:inet_backend` is only valid when specified as the
first argument to `:gen_tcp.listen/2`, this results in a `:badarg`
error.