rewrite integration-test websocket client with Mint.WebSocket
This is a rather large refactor commit which is a complete rewrite
of test/support/websocket_client.exs.

For the integration tests, there are only a handful of changes:

- `start_link/3` has been renamed to `connect/3`, and it acts the same
- `send_message/2` and `send_control_frame/3` have been replaced with
  a unified `send/2` which can send any frame
    - this only affects the tests of the websocket client itself, not
      the channel tests

The WebsocketClient fixture has been overhauled. :websocket_client
is a general-purpose websocket client library that uses processes
to hide some of the minutiae of the websocket runtime - namely
responding to pings with pongs and terminating the process when
a close frame is received.

Mint.WebSocket is functional, so it is less "a websocket client"
and more "tools for writing yourself a websocket client." As such,
the fixture is now a regular GenServer which responds to calls,
casts and info.