fix expected port in transport URI connect-info extraction test
This is necessary because `:websocket_client` sets the HTTP/1
Host header in the bootstrap GET request without the port.

For example, the line 133 test in websocket_socket_test.exs
("returns params with sync request") gives the following request
with `:websocket_client`:

    GET /ws/websocket?key=value HTTP/1.1
    Host: 127.0.0.1
    Connection: Upgrade
    Sec-WebSocket-Version: 13
    Sec-WebSocket-Key: Wd76OpgxB6G3HhApUbE9ig==
    Upgrade: websocket

A run of the same test case with Mint:

    GET /ws/websocket?key=value HTTP/1.1
    host: 127.0.0.1:5907
    user-agent: mint/1.4.1
    upgrade: websocket
    connection: upgrade
    sec-websocket-version: 13
    sec-websocket-key: +fVzxGC42VwPgi2Tm/4bRA==

Note that the host header contains the port. So we update
the test to expect the extracted connection information to
contain the server's port as the client knows it.