Adjust error message
Since Elixir has got the runtime config, it's not necessarily the case that the manifest will be configured in "config/prod.exs", it might be done in the "config/runtime.exs" using env variables for example. For that reason the wording in the error message becomes a bit confusing, I tried to change it as good as I can, sill might need to be more general.