Add missing gettext error messages to pot file
Not all translations in errors.po appear in the template errors.pot.
These are added in errors.pot:

- should be %{count} byte(s)
- should be at least %{count} byte(s)
- should be at most %{count} byte(s)

This avoids these messages being removed when running
`mix gettext.extract --merge` on a clean project.

The order is also consistent, again to avoid changes
when running gettext on a clean project.

Closes #4849