[Socket] send connect error to client and close
Previously if socket connect handler returns an error
then websocket upgrade would be rejected.
Browser websocket api doesn't support getting the status code on upgrade
failure.
This instead successfully upgrades to websocket then sends the error as
the first message and immediately closes the socket.
This is done instead of sending the error in the close frame payload
because that payload is limited to 123 bytes which won't work well for
errors provided by the application developer.

Close status code 3000 is chosen because RFC6455 states
> Status codes in the range 3000-3999 are reserved for use by
> libraries, frameworks, and applications.

This change allows the application developer to return a custom error on
socket connect to indicate for example an authentication/authorization
error.
The client can then implement custom behaviour based on the error.