Add reference to Components guide
In the `A new view` section of the `Request life-cycle`, the reader was being taken directly to the `Phoenix.Component` API documentation.

However, there is now a separate guide for components and HEEx templates (which then links to the API docs).

This links the reader to the components guide instead of the API documentation.