Fixes flaky test case
Test cases in phx.gen.json_test.exs are isolated given gen tasks are run in
different tmp directories, however it seems this particular recently-added test case
having a run time compilation of a PhoenixWeb.CoreComponents module is leaking to
other test case given the module is not purged/deleted.

In particular it seems to affect the test case titled `"with json --web
namespace generates namespaced web modules and directories"` in the same
test file.

So the intermittence in the failure depends on the random order of the
test case run.

Failure can be reproduced consistently with some seeds. E.g.

```
$ mix test test/mix/tasks/phx.gen.json_test.exs --seed 911231
```