Fix race condition in channels during leave on dup topic
This only happens in the longpoll transport due to another race in how
messages are concurrently POST'd to the server.

While we detect duplicate joins and gracefully terminate the
existing channel for the dup topic, we had a race on latent
leaves for an existing topic.

The flow looks like this:

1. channel initiates a phx_close from server
2. xhr POST: client sends a phx_leave before getting the phx_close
3. xhr POST: client immediately sends a join on the same topic
4. transport cleans up original channel
5. phx_join arrives before phx_leave sometimes because the xhr requests
are concurrently handled in the endpoint
6. phx_leave for "old" channel arrives after join and kills new channel

While this race only exists for longpoll, we can fix it by comparing
the join_ref of the current channel with the join_ref of the phx_leave
and ignore the message since the server has already sent the original
channel a phx_close.