Add unique ids flashes in flash_group
The user-provided flashes for :info and :error in the <.flash_group />
component render using the <flash /> component, but do not specify
an id, causing the default id of "flash" to be used.

Since the logic around flashes allow for both info and error to be
present, this may resut in neither rendering, or one rendering over
another.

Providing a unique ID results in things working in a less confusing
way.

This updates the generator for core components to provide this
unique id.