Add --method option to phx.routes
Update the list of valid options in `Mix.Tasks.Phx.Routes`, adding a new
`--method` option, which, when combined with the `--info` option, allows
developers to specify what HTTP method should be used in conjunction
with the specified URL.
This should help in cases where the same URL matches multiple controller
actions, depending on the HTTP method. For example, an application with
a URL such as `/api/v1/users` might use it to list users, with the `GET`
method, or to create a new user, with the `POST` method.