Ensure that the html generator uses correct plural collection
If the plural and singular term are the same when generating a resource,
the template expects a `_collection` value to be set, however this was
only being set by the live generators and not the html generators.

This commect uses the correct `schema.collection` value when generating
a controller.

This closes #5562