Stringify lists in the ChannelTest module (#5576)
When testing with channels, map keys are stringified to replicate what
would happen in the browser. This allows a test such as:

    push(socket, "topic", %{foo: "bar"})

The ChannelTest module will modify this to reflect a real client,
resulting in the channel receiving:

    %{"foo" => "bar"}

This commit also includes arrays, making the behaviour consistent when
pushing data containing an array.