Fix filename when sending downloads with non-ascii names
according to https://www.rfc-editor.org/rfc/rfc6266#appendix-D
when it is not possible to express a filename with regular encoding, a
`filename*` parameter can be added to the header.

This commit adds the `filename*` when the encoded filename does not
match the filename.

This has been tested on chrome, safari and firefox.