Skip map inputs in generated forms
Previously when using the html and live generators, generating a map
field would result in a text input field. There is no trivial way to
input map data in a text field. The map input is now skipped when
generating the html and live forms.

This introduces a new issue where the field is missing, but marked as
required on the schema. To fix this, a new `optionals` value is
available on the generator schema, maps are automatically included into
this schema. Any optionals are now skipped when generating
`validate_required` in the schema file.