Allow UTC datetimes in the generators
The generators now consider the :timestamp_type value in the generator
config for an application.

For backwards compatability, if this value is not set, it will default
to `:naive_datetime` which will generate the timestamps without an
explicit type, otherwise the type will be set explicitly, such as:

`timestamps(:type, :utc_datetime)`

The `mix phx.new` generators also set this value by default, so that
projects going forward default to `:utc_datetime` timestamps.