Automatically migrate when release starts using sqlite 3
When using the sqlite3 ecto adapter, the new and umbrella generators
will add the `Ecto.Migrator` to run the migrations.

By default these run only when inside of a release, but the intent is to
allow the end user to customize this in a way that suits their
production environment.