Add tests for the repo and migration-dir options in schema generator
The `repo` and `migration-dir` options change the location of the
migrations directory, either to the specified directory or one that
matches the repo name (underscored).

The default is explicitly set to "repo" if neither option is set, to
ensure backwards compatability even if the repo has a different name
than the typical `MyApp.Repo`.