Fix phx.routes task for multiple annotations
The `phx.routes` mix task relies on `Code.fetch_docs/1` to find the line
number for a specific function in a module's file. However, it assumes
that `Code.fetch_docs/1` will always return the line number as the
second element of the tuple, which is not always the case, as
`Code.fetch_docs/1` can return a list of annotations as a Keyword list.
When that's the case, the line number is actually the value for the
`:location` key.
As such, this commit updates the code in `phx.routes` to make sure that
both situations are supported.