Add missing gettext calls in core_components.ex
Depending on the context of the template file, we need to output code
with different syntaxes. To make maintenance easier and keep the
with/without Gettext strings consistent, we add a helper
`maybe_gettext/3` to contextually call gettext or skip it, while
outputting correct syntax. The helper is internal and is never exposed
to the generated code.

Both installer and Phoenix (through phx.gen.live) have a
core_components.ex template file, and those must stay in sync. We don't
want to create a dependency between Phoenix and the installer, so
`maybe_gettext/3` is implemented in both places and, similar to the
core_components.ex template itself, we ensure that `maybe_gettext/3`
implementations stay in sync with automated tests.