input validation errors must disappear when resolved
I came across an interesting a corner case.

Validation errors for an input field would never disappear, even after user amends their input to be valid. I noticed that heex template would be re-rendered on the server (`input()` method was called), but UI would never remove the last errors from the view.
   
Getting rid of `defaults: []` fixed this. It seems this is safe to do as errors are always assigned with 

```
    |> assign(:errors, Enum.map(field.errors, &translate_error(&1)))
````