Correct migration step in phx.gen.auth docs
After running `mix phx.gen.auth` you are prompted to run `mix deps.get` followed by `mix ecto.migrate`.

The docs reference `mix ecto.setup` however, which seems to be deprecated in favor of `mix ecto.create`, neither of which are correct in this context, assuming a basic Phoenix app is up and running prior to adding auth.

This PR corrects the docs, replacing `mix ecto.setup` with `mix ecto.migrate` along with an appropriate description of the step.

It also adds the assumption that the reader has gone through the introductory guides and has a running Phoenix app, as well as calling out the addition of the `Register` and `Log in` links added to the default page. 

I also took the liberty to remove some extraneous language about umbrella app root locations, which seemed unnecessary in the context of learning auth basics.