Fix security issue on update_user_email.
During `update_user_email` we have to clean up not only chagne email token,
but also all tokens that were send to email previously.
Such as "confirm" and "reset_password" tokens.
This is because they were sent to previous email,
which potentially can be compromised at this point,
as it can be one of reasons to chagne email initially.