run endpoint config warmup before starting socket (#5985)
There was a slight time window in between starting the socket listener
and actually setting the config in persistent_term that could rely in a
crash on incoming requests. This commit fixes this by calling the warmup
function as a child in the endpoint supervisor, after the config is started,
but before the socket listener is started.

Fixes https://github.com/phoenixframework/phoenix/issues/5981.