Wrap phoenix.js transports
Phoenix expected a transport to be API compatible with the JS
WebSocket API. This is a little bit unflexible and configuring
the authToken that is handled differently depending on the transport
was a bit awkward. Thus, we introduce a new Transport class that
should be implemented by the given transport, falling back to a
a WebSocket compatible wrapper, keeping API compatibility.

The websocket specific handling (readyState, onopen, etc.) could
be further abstracted in the future, to make integrating new
transports like WebTransport easier. Then, it would also make
sense to document the Transport API for others to implement
their own transports.