Prevent duplicating core components between phoenix and installer (#6126)
This commit copies the core_components.ex file from the phoenix
application into the installer when calling `compile` or `hex.publish` .
This file is now `.gitignore` so that it isn't commited to version
control.

The file has also explicitly been added to the list of files in
`mix.exs` to ensure it isn't missing when publishing the package.