Update Routing Cheatsheet/Guide focused on Verified Routes
When the Cheatsheet was created, it got feedback from the Phoenix team
to focus on Verified Routes instead of Router Helpers
(https://github.com/phoenixframework/phoenix/pull/5605#issuecomment-1779497054).

This takes the guide one step further by removing the usage of `as:
:some_name` from the `scope` and `resources` macros, which was
previously used to generate router helpers.

It also updates the examples to showcase how Verified Routes are
typically used, by interpolating variables/assigns directly into the
path.

Finally, it removes the `as: :some_name` option from the nested `scope`
calls in the Routing guide, as it was also not relevant in that context,
because the guide is focused on Verified Routes usage.