Remove reference to OTP version affected by CVE-2025-32433 (#6418)
* Remove reference to OTP version affected by CVE-2025-32433

The example Dockerfile in the "Deploying with Releases" guide references
OTP version 25.3.2.15, which is affected by
[CVE-2025-32433](https://github.com/erlang/otp/security/advisories/GHSA-37cp-fgq5-7wc2).

This is lower than the minimum recommended version of OTP-25.3.2.20.

This pull request changes the values in the example Docker to the use a
newer patched version of the same major version of OTP, which references
an existing image on Docker Hub.

* Bump versions used in example Dockerfile