Support PORT env var using runtime.exs (#6449)
The config/dev.exs file is for build time configuration. This commit
updates the support for the PORT environment variable in new projects
generated with `phx.new` to be done in runtime.exs instead.

The HTTP port is configured regardless of the environment, making it
easier to reason about.

This change is motivated by a desire to not promote the use of
`System.get_env` in config files other than `runtime.exs`, since the
Phoenix templates are often referenced and expected to "teach the right
patterns".

See also
https://github.com/phoenixframework/phoenix/commit/feef6076094104e45660637e05629f5869d6d836.