Fix LongPoll transport name in sessionStorage and logs
Projects created with `mix phx.new` enable JavaScript minification by
default in production. This renames symbols to reduce code size,
including the LongPoll transport class name.

As a result, when the Socket falls back to LongPoll, it stores something
like `phx:fallback:Jt` in sessionStorage instead of the expected
`phx:fallback:LongPoll` (as it does in development).

The shortened name is not guaranteed to be stable across builds, leading
to unintended behavior when deploying new versions of an app.

The transport name was also used in log messages, which now will also
preserve the correct LongPoll name.